+++ apps/playground/src/pages/cientos/controls/TransformControlsDemo.vue
@@ -72,7 +72,6 @@ const { mode, enabled, space, axis, size, showX, showY, showZ } = useControls({
   <TresLeches :uuid="uuid" />
   <TresCanvas
     v-bind="gl"
-    ref="context"
   >
     <TresPerspectiveCamera :position="[11, 11, 11]" :look-at="[0, 0, 0]" />
     <OrbitControls make-default />