+++ apps/playground/src/pages/core/basic/sizes/index.vue
@@ -4,7 +4,7 @@ import { TresCanvas } from '@tresjs/core'
 import { OrbitControls } from '@tresjs/cientos'
 import Child from './Child.vue'
 
-const uuid = 'core-basic-sizes'
+const _uuid = 'core-basic-sizes'
 const state = reactive({
   width: 0,
   height: 0,